Fixed ambiguous overload for std::pow() error.
Calling pow(int,int) is ambiguous with gcc 4.1 compilers, which caused
numerous build errors in the nightly tests.  I added explicit casts
to double to fix this.  Note that the result is an int, so casting
to value_type doesn't make sense if value_type is not a scalar,
so I used double.

+++ packages/stokhos/src/Stokhos_KL_ExponentialRandomFieldImp.hpp
@@ -61,7 +61,8 @@ ExponentialRandomField(Teuchos::ParameterList& solverParams)
   }
 
   // Compute all possible tensor product combinations of 1-D eigenfunctions
-  int num_prod = std::pow(num_KL, dim);
+  int num_prod = static_cast<int>(std::pow(static_cast<double>(num_KL), 
+					   static_cast<double>(dim)));
   product_eig_pairs.resize(num_prod);
   Teuchos::Array<int> index(dim, 0);
   int cnt = 0;
